Update public docs.
