+++ lib/src/cli/commands/create_command.dart
@@ -1,13 +1,7 @@
-import 'dart:io';
-
 import 'package:args/args.dart';
-import 'package:path/path.dart' as p;
-import 'package:sprintf/sprintf.dart';
-import 'package:yaml/yaml.dart';
 
 import '../flag_commands/flag_commands.dart';
 import '../models/command_names.dart';
-import '../models/default_config_templates.dart';
 import '../models/flag_names.dart';
 import '../utils/utils.dart';
 import 'base_command.dart';
@@ -25,6 +19,8 @@ class CreateCommand extends BaseCommand {
   @override
   String get summary => 'Creates config file in the root of the project.';
 
+  /// Default constructor for create command.
+  /// [logger] is used to output all kinds of logs, errors and exceptions.
   CreateCommand(super.logger) {
     argParser
       ..addFlag(FlagNames.addInPubspec,
@@ -58,106 +54,5 @@ class CreateCommand extends BaseCommand {
       if (result.stacktrace != null) verbose(result.stacktrace.toString());
       exitWith(result.error);
     }
-  }
-}
-
-class ConfigCreator {
-  final BaseLogger? logger;
-
-  ConfigCreator([this.logger]);
-
-  Result<void> create({
-    bool addInPubspec = false,
-    bool isJson = false,
-    String? path,
-  }) {
-    try {
-      if (addInPubspec) {
-        final result = createConfigsInPubspec();
-        if (result.isError) return result;
-        return Result.success();
-      }
-      if (path != null && path.isNotEmpty) {
-        final result = createConfigFileAtCustomPath(path, isJson);
-        if (result.isError) return result;
-        return Result.success();
-      }
-      final result = createConfigFileInCurrentDirectory(isJson);
-      if (result.isError) return result;
-      return Result.success();
-    } on Error catch (error, stacktrace) {
-      return Result.error('Unable to create config file', error, stacktrace);
-    }
-  }
-
-  Result<void> createConfigsInPubspec() {
-    final pubspecFile = file(p.join(Directory.current.path, 'pubspec.yaml')) ??
-        file(p.join(Directory.current.path, 'pubspec.yml'));
-    if (pubspecFile == null) {
-      return Result.error(ConsoleMessages.pubspecNotFound);
-    }
-    final pubspecContent = pubspecFile.readAsStringSync();
-    final pubspec = loadYaml(pubspecContent);
-    if (pubspec['spider'] != null) {
-      return Result.error(ConsoleMessages.configExistsInPubspec);
-    }
-    try {
-      final lines = pubspecFile.readAsLinesSync();
-      String configContent = DefaultConfigTemplates.pubspecFormat;
-      if (lines.last.trim().isNotEmpty || !lines.last.endsWith('\n')) {
-        configContent = '\n\n$configContent';
-      }
-      pubspecFile.openWrite(mode: FileMode.writeOnlyAppend)
-        ..write(configContent)
-        ..close();
-      logger?.success(ConsoleMessages.configCreatedInPubspec);
-      return Result.success();
-    } on Error catch (error, stacktrace) {
-      return Result.error(
-          ConsoleMessages.unableToAddConfigInPubspec, error, stacktrace);
-    }
-  }
-
-  /// Creates config file at custom path.
-  Result<void> createConfigFileAtCustomPath(String path, bool isJson) {
-    String filePath;
-    if (path.startsWith('/')) path = '.$path';
-    if (FileSystemEntity.isDirectorySync(path)) {
-      // provided path is an existing directory
-      Directory(path).createSync(recursive: true);
-      filePath = p.join(path, isJson ? 'spider.json' : 'spider.yaml');
-    } else {
-      final String extension = p.extension(path);
-      if (extension.isNotEmpty) {
-        return Result.error('Provided path is not a valid directory.');
-      }
-      Directory(path).createSync(recursive: true);
-      filePath = p.join(path, isJson ? 'spider.json' : 'spider.yaml');
-    }
-    final content = p.extension(filePath) == '.json'
-        ? DefaultConfigTemplates.jsonFormat
-        : DefaultConfigTemplates.yamlFormat;
-    final file = File(filePath);
-    if (file.existsSync()) {
-      return Result.error('Config file already exists at $filePath.');
-    }
-    file.writeAsStringSync(content);
-    logger
-        ?.success(sprintf(ConsoleMessages.fileCreatedAtCustomPath, [filePath]));
-    return Result.success();
-  }
-
-  Result<void> createConfigFileInCurrentDirectory(bool isJson) {
-    final filename = isJson ? 'spider.json' : 'spider.yaml';
-    final dest = File(p.join(Directory.current.path, filename));
-    final content = isJson
-        ? DefaultConfigTemplates.jsonFormat
-        : DefaultConfigTemplates.yamlFormat;
-    if (dest.existsSync()) {
-      logger?.info('Config file already exists. Overwriting configs...');
-    }
-    dest.writeAsStringSync(content);
-    logger?.success('Configuration file created successfully.');
-    return Result.success();
   }
 }
