+++ lib/src/cli/flag_commands/license_flag_command.dart
@@ -17,6 +17,7 @@ class LicenseFlagCommand extends BaseFlagCommand {
   @override
   bool get negatable => false;
 
+  /// Default constructor.
   LicenseFlagCommand(super.logger);
 
   @override
