+++ lib/src/cli/models/asset_group.dart
@@ -23,16 +23,34 @@ import 'asset_subgroup.dart';
 
 /// Holds group information for assets sub directories
 class AssetGroup {
+  /// Name of the dart class for the group.
   late final String className;
+
+  /// name of the dart file for the group.
   late final String fileName;
+
+  /// Whether to use underscore over camel case for the reference names.
   late final bool useUnderScores;
+
+  /// Whether to generate static references or not.
   late final bool useStatic;
+
+  /// Whether to generate constant references or not.
   late final bool useConst;
+
+  /// Subgroups of the group.
   late final List<AssetSubgroup>? subgroups;
+
+  /// Whitelisted file types for the group.
   late final List<String>? types;
+
+  /// Paths covered by the group.
   late final List<String>? paths;
+
+  /// Prefix to append to the reference names.
   late final String? prefix;
 
+  /// Default constructor.
   AssetGroup({
     required this.className,
     String? fileName,
@@ -41,6 +59,7 @@ class AssetGroup {
     this.useConst = true,
   }) : fileName = Formatter.formatFileName(fileName ?? className);
 
+  /// Generates [AssetGroup] from the [json] map data.
   AssetGroup.fromJson(Map<String, dynamic> json) {
     className = json['class_name'].toString();
     fileName =
