+++ lib/src/cli/models/asset_subgroup.dart
@@ -2,16 +2,23 @@ import '../../generation_utils.dart';
 
 /// Holds prefix, types and path info for [Group]
 class AssetSubgroup {
+  /// Prefix to be appended to the reference names.
   late final String prefix;
+
+  /// Whitelisted types that are part of this sub-group.
   late final List<String> types;
+
+  /// Paths covered under this sub-group.
   late final List<String> paths;
 
+  /// Default constructor.
   AssetSubgroup({
     required this.paths,
     this.prefix = '',
     this.types = const <String>[],
   });
 
+  /// Creates [AssetSubgroup] from [json] map data.
   AssetSubgroup.fromJson(Map<String, dynamic> json) {
     prefix = json['prefix']?.toString() ?? '';
     types = <String>[];
