+++ lib/src/cli/models/command_names.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: public_member_api_docs
+
 /// Contains names of all commands available in the CLI.
 class CommandNames {
   CommandNames._();
