+++ lib/src/cli/utils/utils.dart
@@ -1,3 +1,4 @@
+export 'config_creator.dart';
 export 'config_retriever.dart';
 export 'constants.dart';
 export 'extensions.dart';
