+++ lib/src/data/yaml_config_template.dart
@@ -44,6 +44,7 @@ groups:
 #    package: res
 #    types: [ .svg ]''';
 
+/// A template for config creation inside pubspec.yaml file.
 const String YAML_PUBSPEC_CONFIG_TEMPLATE = '''
 # Generated by Spider
 spider:
